move parts of code to modules
