+++ js/module/module-artist.js
@@ -1,19 +1,11 @@
 // Выбор исполнителя: уровень
 import getElementFromTemplate from '../utils/getElementFromTemplate';
-// import showScreen from './../utils/showScreen';
-// import genreScreen from './module-genre';
-// import resultScreen from './module-result';
 import artistList from '../components/artist-list';
 import timer from '../components/timer';
 import game from '../data/game';
-// import lossResult from '../data/result/loss';
 import player from '../components/player';
-// import genre from '../data/genre';
-import questionList from '../data/questions-list';
-import switchNextScreen from '../utils/switchNextScreen';
-import switchQuestion from '../utils/switchQuestion';
-import reduceLives from '../utils/reduceLives';
-import isQuestionOver from './../utils/isQuestionOver';
+import doOnSuccess from '../utils/doOnSuccess';
+import doOnFault from '../utils/doOnFault';
 
 export default (data) => {
   const artistTemplate = `
@@ -40,25 +32,10 @@ export default (data) => {
     if (evt.target.classList.contains(`main-answer-r`)) {
       if (data.trackArtist.toLowerCase() === evt.target.getAttribute(`value`)) {
 
-        const currentQuestion = window.sessionStorage.getItem(`currentQuestion`);
-        const nextQuestion = switchQuestion(currentQuestion);
-        window.sessionStorage.setItem(`currentQuestion`, nextQuestion);
-        const result = isQuestionOver(currentQuestion) ?
-          `win` : questionList[nextQuestion].type;
+        doOnSuccess();
 
-        switchNextScreen(result);
       } else {
-        const currentLives = reduceLives(window.sessionStorage.getItem(`numberOfLive`));
-
-        if (currentLives) {
-          window.sessionStorage.setItem(`numberOfLive`, currentLives);
-
-          const nextQuestion = switchQuestion(window.sessionStorage.getItem(`currentQuestion`));
-          window.sessionStorage.setItem(`currentQuestion`, nextQuestion);
-          switchNextScreen(questionList[nextQuestion].type);
-        } else {
-          switchNextScreen(`loss`);
-        }
+        doOnFault();
       }
     }
   });
