+++ js/module/module-genre.js
@@ -1,16 +1,9 @@
 // Игра на выбор жанра
 import getElementFromTemplate from '../utils/getElementFromTemplate';
-// import showScreen from './../utils/showScreen';
-// import winResult from '../data/result/win';
-// import lossResult from '../data/result/loss';
 import genreList from '../components/genre-list';
-// import result from './module-result';
 import isAllCheckedCorrectly from '../utils/check-genre';
-import isQuestionOver from './../utils/isQuestionOver';
-import switchNextScreen from './../utils/switchNextScreen';
-import switchQuestion from './../utils/switchQuestion';
-import questionList from '../data/questions-list';
-import reduceLives from '../utils/reduceLives';
+import doOnSuccess from '../utils/doOnSuccess';
+import doOnFault from '../utils/doOnFault';
 
 export default (data) => {
   const genre = getElementFromTemplate(`
@@ -45,25 +38,11 @@ export default (data) => {
     evt.preventDefault();
 
     if (isAllCheckedCorrectly(inputs, data)) {
-      const currentQuestion = window.sessionStorage.getItem(`currentQuestion`);
-      const nextQuestion = switchQuestion(currentQuestion);
-      window.sessionStorage.setItem(`currentQuestion`, nextQuestion);
-      const result = isQuestionOver(currentQuestion) ?
-        `win` : questionList[nextQuestion].type;
 
-      switchNextScreen(result);
+      doOnSuccess();
+
     } else {
-      const currentLives = reduceLives(window.sessionStorage.getItem(`numberOfLive`));
-
-      if (currentLives) {
-        window.sessionStorage.setItem(`numberOfLive`, currentLives);
-
-        const nextQuestion = switchQuestion(window.sessionStorage.getItem(`currentQuestion`));
-        window.sessionStorage.setItem(`currentQuestion`, nextQuestion);
-        switchNextScreen(questionList[nextQuestion].type);
-      } else {
-        switchNextScreen(`loss`);
-      }
+      doOnFault();
     }
   });
